refactor: migrate to Pages Router by removing App Router components and updating navigation hooks
