+++ src/contexts/AuthContext.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { createContext, useContext, useEffect, useState, type ReactNode } from "react";
 import { User, Session } from "@supabase/supabase-js";
 import { supabase } from "@/lib/supabaseClient";
